pyproject.toml
[tool.poetry]
name = "probabilistic-timed-automata"
version = "0.1.0"
description = "Probabilistic Timed Automata library for Python"
authors = ["Anand Balakrishnan <anandbala1597@gmail.com>"]

license = "MIT"

readme = "README.md"
repository = "https://github.com/anand-bala/probabilistic-timed-automata.git"

packages = [
  { include = "pta" },
]

[tool.poetry.dependencies]
python = "^3.6"
attrs = "^19.3.0"
portion = "^2.0.0"
sphinx = {version = "^2.4.4", allow-prereleases = true}

[tool.poetry.dev-dependencies]
pytest = "^5.2"
mypy = "^0.761"
black = "^19.10b0"
Sphinx = "^2.4.4"
sphinx-better-theme = "^0.1.5"
autoflake = "^1.3.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
